requirements.txt
spotipy==2.26.0
python-dotenv==1.2.2
fastapi
uvicorn[standard]
httpx
psycopg2-binary
